Moved GLSLCompileException
