+++ src/main/java/jgloom/common/gl/glsl/GLSLCompileException.java
@@ -1,5 +1,6 @@
-package jgloom;
+package jgloom.common.gl.glsl;
 
+import jgloom.GLNativeException;
 import jgloom.gl.glsl.GLSLShader;
 
 /**
