+++ src/main/java/jgloom/common/gl/glsl/GLSLShaderContainer.java
@@ -3,7 +3,6 @@ package jgloom.common.gl.glsl;
 import org.lwjgl.opengl.GL11;
 import org.lwjgl.opengl.GL20;
 
-import jgloom.GLSLCompileException;
 import jgloom.gl.glsl.GLSLShader;
 
 /**
